requirements (4).txt
streamlit
transformers
PyMuPDF
python-docx
torch
tensorflow
